Set up abstraction flow to use error-reporter, allowing the Blender plugin to easily report errors when they happen.
