+++ yk_gmd_blender/blender/importer/legacy/importer.py
