+++ yk_gmd_blender/yk_gmd/v2/io.py
@@ -2,20 +2,25 @@ from pathlib import Path
 from typing import Union, Tuple, cast
 
 from yk_gmd_blender.yk_gmd.v2.abstract.gmd_scene import GMDScene
+from yk_gmd_blender.yk_gmd.v2.converters.yk1.to_abstract import GMDAbstractor_YK1
 from yk_gmd_blender.yk_gmd.v2.errors.error_classes import InvalidGMDFormatError
+from yk_gmd_blender.yk_gmd.v2.errors.error_reporter import ErrorReporter
+from yk_gmd_blender.yk_gmd.v2.structure.common.file import FileUnpackError
 from yk_gmd_blender.yk_gmd.v2.structure.common.header import GMDHeaderStruct, GMDHeaderStruct_Unpack
 from yk_gmd_blender.yk_gmd.v2.structure.endianness import check_is_file_big_endian
 from yk_gmd_blender.yk_gmd.v2.structure.kenzan.file import FileData_Kenzan, FilePacker_Kenzan
 from yk_gmd_blender.yk_gmd.v2.structure.version import GMDVersion, VersionProperties
-from yk_gmd_blender.yk_gmd.v2.converters.yk1.from_abstract import read_abstract_contents_YK1
 from yk_gmd_blender.yk_gmd.v2.structure.yk1.file import FileData_YK1, FilePacker_YK1
 
 
-def _get_file_data(data: Union[Path, str, bytes]) -> bytes:
+def _get_file_data(data: Union[Path, str, bytes], error_reporter: ErrorReporter) -> bytes:
     if isinstance(data, (Path, str)):
-        with open(data, "rb") as in_file:
-            data = in_file.read()
-        return data
+        try:
+            with open(data, "rb") as in_file:
+                data = in_file.read()
+            return data
+        except FileNotFoundError as e:
+            error_reporter.fatal(str(e))
     else:
         return data
 
@@ -29,31 +34,38 @@ def _extract_base_header(data: bytes) -> Tuple[bool, GMDHeaderStruct]:
     return big_endian, base_header
 
 
-def read_gmd_structures(data: Union[Path, str, bytes]) -> Tuple[VersionProperties, Union[FileData_Kenzan, FileData_YK1]]:
-    data = _get_file_data(data)
+def read_gmd_structures(data: Union[Path, str, bytes], error_reporter: ErrorReporter) -> Tuple[VersionProperties, Union[FileData_Kenzan, FileData_YK1]]:
+    data = _get_file_data(data, error_reporter)
     big_endian, base_header = _extract_base_header(data)
 
     version_props = base_header.get_version_properties()
     if version_props.major_version == GMDVersion.Kiwami1:
-        contents, _ = FilePacker_YK1.unpack(big_endian, data=data, offset=0)
+        try:
+            contents, _ = FilePacker_YK1.unpack(big_endian, data=data, offset=0)
 
-        return version_props, contents
+            return version_props, contents
+        except FileUnpackError as e:
+            error_reporter.fatal(str(e))
     elif version_props.major_version == GMDVersion.Kenzan:
-        contents, _ = FilePacker_Kenzan.unpack(big_endian, data=data, offset=0)
+        try:
+            contents, _ = FilePacker_Kenzan.unpack(big_endian, data=data, offset=0)
 
-        return version_props, contents
+            return version_props, contents
+        except FileUnpackError as e:
+            error_reporter.fatal(str(e))
     else:
         raise InvalidGMDFormatError(f"File format version {version_props.version_str} is not readable")
 
 
-def read_abstract_scene_from_contents(version_props: VersionProperties, contents: Union[FileData_Kenzan, FileData_YK1]) -> GMDScene:
+def read_abstract_scene_from_filedata_object(version_props: VersionProperties, contents: Union[FileData_Kenzan, FileData_YK1], error_reporter: ErrorReporter) -> GMDScene:
     if version_props.major_version == GMDVersion.Kiwami1:
-        return read_abstract_contents_YK1(version_props, cast(FileData_YK1, contents))
+        return GMDAbstractor_YK1(version_props, cast(FileData_YK1, contents), error_reporter).make_abstract_scene()
     else:
         raise InvalidGMDFormatError(f"File format version {version_props.version_str} is not abstractable")
 
-def read_abstract_scene_from_bytes(data: Union[Path, str, bytes]) -> GMDScene:
-    data = _get_file_data(data)
 
-    version_props, file_data = read_gmd_structures(data)
-    return read_abstract_scene_from_contents(version_props, file_data)
+def read_abstract_scene(data: Union[Path, str, bytes], error_reporter: ErrorReporter) -> GMDScene:
+    data = _get_file_data(data, error_reporter)
+
+    version_props, file_data = read_gmd_structures(data, error_reporter)
+    return read_abstract_scene_from_filedata_object(version_props, file_data, error_reporter)
